requirements.txt

# Python 3.11 recommended

# API
fastapi==0.115.14
uvicorn==0.30.6
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
pydantic-core==2.33.2
starlette==0.46.2

# LangChain
langchain==0.2.17
langchain-community==0.2.19
langchain-openai==0.1.25
langchain-huggingface==0.0.3
openai==1.106.1

# langgraph
langgraph==0.2.76

# Vector store
faiss-cpu==1.8.0.post1

# HF fallback
transformers==4.56.1
accelerate==0.34.2
sentencepiece==0.2.1
safetensors==0.4.5
sentence-transformers==5.1.0
tokenizers==0.22.0

